+++ app/src/main/java/org/linphone/utils/DataBindingUtils.kt
@@ -34,6 +34,8 @@ import androidx.appcompat.content.res.AppCompatResources
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.constraintlayout.widget.Guideline
 import androidx.databinding.*
+import androidx.emoji2.emojipicker.EmojiPickerView
+import androidx.emoji2.emojipicker.EmojiViewItem
 import coil.dispose
 import coil.load
 import coil.request.CachePolicy
@@ -791,4 +793,15 @@ fun ImageView.setPresenceIcon(presence: ConsolidatedPresence) {
         else -> AppUtils.getString(R.string.content_description_presence_offline)
     }
     setContentDescription(contentDescription)
+}
+
+interface EmojiPickedListener {
+    fun onEmojiPicked(item: EmojiViewItem)
+}
+
+@BindingAdapter("emojiPickedListener")
+fun EmojiPickerView.setEmojiPickedListener(listener: EmojiPickedListener) {
+    setOnEmojiPickedListener { emoji ->
+        listener.onEmojiPicked(emoji)
+    }
 }